requirements.txt
opencv-python~=4.5.5.64
numpy~=1.22.3
PyQt5~=5.15.6
Pillow~=9.0.1
pip~=22.0.4
cfbd~=4.3.2
six~=1.16.0
certifi~=2021.10.8
urllib3~=1.26.6
pytz~=2021.3
requests~=2.27.1
setuptools~=60.9.3
pandas~=1.4.1
matplotlib~=3.5.1
python-dateutil~=2.8.2
chardet~=4.0.0
psycopg2~=2.9.3
idna~=2.10
roboflow~=0.2.2
wget~=3.2
tqdm~=4.63.0
python-dotenv~=0.19.2
fonttools~=4.30.0
pyparsing~=2.4.7
cycler~=0.10.0
packaging~=21.3
kiwisolver~=1.3.1